Build and verify AgonDev MOS candidate

+++ projects/mos-port/include/gpio.h
@@ -1,9 +1,9 @@
 #ifndef MOS_AGONDEV_GPIO_COMPAT_H
 #define MOS_AGONDEV_GPIO_COMPAT_H
 
-#include <eZ80.h>
+#include <ez80.h>
 
-#define SETREG(reg, bits) ((reg) |= (bits))
-#define RESETREG(reg, bits) ((reg) &= (uint8_t)~(bits))
+#define SETREG(reg, bits) ((reg) |= (uint8_t)(bits))
+#define RESETREG(reg, bits) ((reg) &= (uint8_t)~(uint8_t)(bits))
 
 #endif
